Migrate SLAMD source from subversion to git
A commit of the SLAMD source code checked out from a subversion
repository.  I'm not going to try to port the subversion commit
history because it's actually not that extensive (SLAMD was
originally hosted in a CVS repository before it was migrated to
subversion, and the history was not preserved as part of that
migration), and because I'm excluding an obsolete "www" top-level
directory that was part of the java.net hosting infrastructure that
really isn't necessary here.

I haven't made any changes at all to the SLAMD source code.  The
code committed here is how it appeared after the last subversion
commit on March 20, 2010.  The dependencies are all horribly out
of date, and there are tons of build warnings when compiling on
Java 8.  However, it seems to run just fine (or at least, as well
as it did when I last stopped working on it).