Update the Berkeley DB Java Edition
Updated the Berkeley DB Java Edition libraries used by SLAMD from
version 3.3.75 (built in November of 2008) to version 7.5.11 (built
in October of 2017).  The license for the Berkeley DB Java Edition
has also changed, from its own thing to Apache license version 2.0.

Note that some minor code changes were required by this update.
In particular, JE's DatabaseException used to be a concrete class
that could be directly instantiated.  It is now abstract.  To work
around this, I have created a new SLAMDDatabaseException subclass,
and updated the code that previously created a DatabaseException so
that it now creates a SLAMDDatabaseException.