Make SLAMD build cleanly on Java 8
Updated tons of source files to make SLAMD build without any
warnings on Java 8.  This was pretty tedious, as SLAMD was
originally created in the days before generics, and a lot of the
code was still using raw types.  Also, some of the APIs it was
using had been deprecated in the last decade, so those have also
been updated to use the preferred non-deprecated replacements.