Add a Basic Search Rate job
Added a Basic Search Rate job that can repeatedly search an LDAP
directory server with a basic set of parameters.

The existing Search Rate job has been moved to the legacy package.
An Advanced Search Rate job should soon be created to take its
place with most of the same options, as well as additional
capabilities.