Add SLAMD server start and stop scripts
Added start-slamd-server.sh and stop-slamd-server.sh shell scripts
in the SLAMD server root directory that make it more obvious how to
start and stop the SLAMD server.  These scripts simply invoke the
Tomcat startup.sh and shutdown.sh scripts in the bin directory, but
the new scripts are easier to find because they're in the top-level
directory, and because it's more obvious what they do.

There are also corresponding start-slamd-server.bat and
stop-slamd-server.bat batch files for use on Windows.