Make SLAMD build cleanly on Java 12
Updated SLAMD so that it builds cleanly with no code warnings on
Java 12.  SLAMD requires Java 8 or later, and it already built
cleanly on Java 8, but it generated a number of deprecation warnings
when building with newer Java versions.  Those warnings have been
addressed, and the code should build cleanly with Java versions 8
through 12.