Enable checkstyle and fix errors it calls out
Updated checkstyle to use the latest version.  Enabled it in the
build process, and fixed a lot of problems that it found.  Note that
some checkstyle rules are currently disabled because there are too
many violations to fix right now.  I hope to address these in the
future, but the set of changes here are good.